refactor the entrypoint dir so each command is its own file
