+++ build_runner/lib/src/entrypoint/base_command.dart
@@ -1,0 +1,92 @@
+// Copyright (c) 2018, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:args/command_runner.dart';
+import 'package:logging/logging.dart';
+
+import '../package_graph/apply_builders.dart';
+import '../package_graph/package_graph.dart';
+import 'options.dart';
+import 'runner.dart';
+
+abstract class BuildRunnerCommand extends Command<int> {
+  Logger get logger => new Logger(name);
+
+  List<BuilderApplication> get builderApplications =>
+      (runner as BuildCommandRunner).builderApplications;
+
+  PackageGraph get packageGraph => (runner as BuildCommandRunner).packageGraph;
+
+  BuildRunnerCommand() {
+    _addBaseFlags();
+  }
+
+  void _addBaseFlags() {
+    argParser
+      ..addFlag(assumeTtyOption,
+          help: 'Enables colors and interactive input when the script does not'
+              ' appear to be running directly in a terminal, for instance when it'
+              ' is a subprocess',
+          negatable: true)
+      ..addFlag(deleteFilesByDefaultOption,
+          help:
+              'By default, the user will be prompted to delete any files which '
+              'already exist but were not known to be generated by this '
+              'specific build script.\n\n'
+              'Enabling this option skips the prompt and deletes the files. '
+              'This should typically be used in continues integration servers '
+              'and tests, but not otherwise.',
+          negatable: false,
+          defaultsTo: false)
+      ..addFlag(lowResourcesModeOption,
+          help: 'Reduce the amount of memory consumed by the build process. '
+              'This will slow down builds but allow them to progress in '
+              'resource constrained environments.',
+          negatable: false,
+          defaultsTo: false)
+      ..addOption(configOption,
+          help: 'Read `build.<name>.yaml` instead of the default `build.yaml`',
+          abbr: 'c')
+      ..addFlag(failOnSevereOption,
+          help: 'Whether to consider the build a failure on an error logged.',
+          negatable: true,
+          defaultsTo: false)
+      ..addFlag(trackPerformanceOption,
+          help: r'Enables performance tracking and the /$perf page.',
+          negatable: true,
+          defaultsTo: false)
+      ..addFlag(skipBuildScriptCheckOption,
+          help: r'Skip validation for the digests of files imported by the '
+              'build script.',
+          hide: true,
+          defaultsTo: false)
+      ..addMultiOption(outputOption,
+          help: 'A directory to write the result of a build to. Or a mapping '
+              'from a top-level directory in the package to the directory to '
+              'write a filtered build output to. For example "web:deploy".',
+          abbr: 'o')
+      ..addFlag(verboseOption,
+          abbr: 'v',
+          defaultsTo: false,
+          negatable: false,
+          help: 'Enables verbose logging.')
+      ..addFlag(releaseOption,
+          abbr: 'r',
+          defaultsTo: false,
+          negatable: true,
+          help: 'Build with release mode defaults for builders.')
+      ..addMultiOption(defineOption,
+          splitCommas: false,
+          help: 'Sets the global `options` config for a builder by key.');
+  }
+
+  /// Must be called inside [run] so that [argResults] is non-null.
+  ///
+  /// You may override this to return more specific options if desired, but they
+  /// must extend [SharedOptions].
+  SharedOptions readOptions() {
+    return new SharedOptions.fromParsedArgs(
+        argResults, argResults.rest, packageGraph.root.name, this);
+  }
+}
