+++ build_runner/lib/src/entrypoint/options.dart
@@ -2,104 +2,33 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'dart:async';
 import 'dart:convert';
 import 'dart:io';
 
 import 'package:args/args.dart';
 import 'package:args/command_runner.dart';
 import 'package:build_config/build_config.dart';
-import 'package:http_multi_server/http_multi_server.dart';
-import 'package:io/io.dart';
-import 'package:logging/logging.dart';
 import 'package:meta/meta.dart';
 import 'package:path/path.dart' as p;
-import 'package:shelf/shelf_io.dart';
 
-import '../asset/file_based.dart';
-import '../asset_graph/graph.dart';
-import '../asset_graph/node.dart';
-import '../generate/build.dart';
-import '../generate/build_result.dart';
-import '../logging/logging.dart';
-import '../logging/std_io_logging.dart';
-import '../package_graph/apply_builders.dart';
-import '../package_graph/package_graph.dart';
-import '../server/server.dart';
-import '../util/constants.dart';
-
-const _assumeTty = 'assume-tty';
-const _define = 'define';
-const _deleteFilesByDefault = 'delete-conflicting-outputs';
-const _logRequests = 'log-requests';
-const _lowResourcesMode = 'low-resources-mode';
-const _failOnSevere = 'fail-on-severe';
-const _hostname = 'hostname';
-const _output = 'output';
-const _config = 'config';
-const _verbose = 'verbose';
-const _release = 'release';
-const _trackPerformance = 'track-performance';
-const _skipBuildScriptCheck = 'skip-build-script-check';
-
-final _pubBinary = Platform.isWindows ? 'pub.bat' : 'pub';
+const assumeTtyOption = 'assume-tty';
+const defineOption = 'define';
+const deleteFilesByDefaultOption = 'delete-conflicting-outputs';
+const logRequestsOption = 'log-requests';
+const lowResourcesModeOption = 'low-resources-mode';
+const failOnSevereOption = 'fail-on-severe';
+const hostnameOption = 'hostname';
+const outputOption = 'output';
+const configOption = 'config';
+const verboseOption = 'verbose';
+const releaseOption = 'release';
+const trackPerformanceOption = 'track-performance';
+const skipBuildScriptCheckOption = 'skip-build-script-check';
 
 final _defaultWebDirs = const ['web', 'test', 'example', 'benchmark'];
 
-/// Unified command runner for all build_runner commands.
-class BuildCommandRunner extends CommandRunner<int> {
-  final List<BuilderApplication> builderApplications;
-
-  final packageGraph = new PackageGraph.forThisPackage();
-
-  BuildCommandRunner(List<BuilderApplication> builderApplications)
-      : this.builderApplications = new List.unmodifiable(builderApplications),
-        super('build_runner', 'Unified interface for running Dart builds.') {
-    addCommand(new _BuildCommand());
-    addCommand(new _WatchCommand());
-    addCommand(new _ServeCommand());
-    addCommand(new _TestCommand());
-    addCommand(new _CleanCommand());
-  }
-
-  // CommandRunner._usageWithoutDescription is private – this is a reasonable
-  // facsimile.
-  /// Returns [usage] with [description] removed from the beginning.
-  String get usageWithoutDescription => LineSplitter
-      .split(usage)
-      .skipWhile((line) => line == description || line.isEmpty)
-      .join('\n');
-}
-
-/// Returns a map of output directory to root input directory to be used
-/// for merging.
-///
-/// Each output option is split on `:` where the first value is the
-/// root input directory and the second value output directory.
-/// If no delimeter is provided the root input directory will be null.
-Map<String, String> _parseOutputMap(ArgResults argResults) {
-  var outputs = argResults[_output] as List<String>;
-  if (outputs == null) return null;
-  var result = <String, String>{};
-  for (var option in argResults[_output] as List<String>) {
-    var split = option.split(':');
-    if (split.length == 1) {
-      var output = split.first;
-      result[output] = null;
-    } else if (split.length >= 2) {
-      var output = split.sublist(1).join(':');
-      var root = split.first;
-      if (root.contains('/')) {
-        throw 'Input root can not be nested: $option';
-      }
-      result[output] = split.first;
-    }
-  }
-  return result;
-}
-
 /// Base options that are shared among all commands.
-class _SharedOptions {
+class SharedOptions {
   /// Skip the `stdioType()` check and assume the output is going to a terminal
   /// and that we can accept input on stdin.
   final bool assumeTty;
@@ -143,7 +72,7 @@ class _SharedOptions {
   /// The directories that should be built.
   final List<String> buildDirs;
 
-  _SharedOptions._({
+  SharedOptions._({
     @required this.assumeTty,
     @required this.deleteFilesByDefault,
     @required this.failOnSevere,
@@ -158,7 +87,7 @@ class _SharedOptions {
     @required this.buildDirs,
   });
 
-  factory _SharedOptions.fromParsedArgs(ArgResults argResults,
+  factory SharedOptions.fromParsedArgs(ArgResults argResults,
       Iterable<String> positionalArgs, String rootPackage, Command command) {
     var outputMap = _parseOutputMap(argResults);
     var buildDirs = _buildDirsFromOutputMap(outputMap);
@@ -172,31 +101,31 @@ class _SharedOptions {
       buildDirs.add(arg);
     }
 
-    return new _SharedOptions._(
-      assumeTty: argResults[_assumeTty] as bool,
-      deleteFilesByDefault: argResults[_deleteFilesByDefault] as bool,
-      failOnSevere: argResults[_failOnSevere] as bool,
-      enableLowResourcesMode: argResults[_lowResourcesMode] as bool,
-      configKey: argResults[_config] as String,
+    return new SharedOptions._(
+      assumeTty: argResults[assumeTtyOption] as bool,
+      deleteFilesByDefault: argResults[deleteFilesByDefaultOption] as bool,
+      failOnSevere: argResults[failOnSevereOption] as bool,
+      enableLowResourcesMode: argResults[lowResourcesModeOption] as bool,
+      configKey: argResults[configOption] as String,
       outputMap: outputMap,
-      trackPerformance: argResults[_trackPerformance] as bool,
-      skipBuildScriptCheck: argResults[_skipBuildScriptCheck] as bool,
-      verbose: argResults[_verbose] as bool,
+      trackPerformance: argResults[trackPerformanceOption] as bool,
+      skipBuildScriptCheck: argResults[skipBuildScriptCheckOption] as bool,
+      verbose: argResults[verboseOption] as bool,
       builderConfigOverrides:
-          _parseBuilderConfigOverrides(argResults[_define], rootPackage),
-      isReleaseBuild: argResults[_release] as bool,
+          _parseBuilderConfigOverrides(argResults[defineOption], rootPackage),
+      isReleaseBuild: argResults[releaseOption] as bool,
       buildDirs: buildDirs.toList(),
     );
   }
 }
 
-/// Options specific to the [_ServeCommand].
-class _ServeOptions extends _SharedOptions {
+/// Options specific to the `serve` command.
+class ServeOptions extends SharedOptions {
   final String hostName;
   final bool logRequests;
-  final List<_ServeTarget> serveTargets;
+  final List<ServeTarget> serveTargets;
 
-  _ServeOptions._({
+  ServeOptions._({
     @required this.hostName,
     @required this.logRequests,
     @required this.serveTargets,
@@ -227,9 +156,9 @@ class _ServeOptions extends _SharedOptions {
           buildDirs: buildDirs,
         );
 
-  factory _ServeOptions.fromParsedArgs(ArgResults argResults,
+  factory ServeOptions.fromParsedArgs(ArgResults argResults,
       Iterable<String> positionalArgs, String rootPackage, Command command) {
-    var serveTargets = <_ServeTarget>[];
+    var serveTargets = <ServeTarget>[];
     var nextDefaultPort = 8080;
     for (var arg in positionalArgs) {
       var parts = arg.split(':');
@@ -245,12 +174,12 @@ class _ServeOptions extends _SharedOptions {
         throw new UsageException(
             'Unable to parse port number in `$arg`', command.usage);
       }
-      serveTargets.add(new _ServeTarget(path, port));
+      serveTargets.add(new ServeTarget(path, port));
     }
     if (serveTargets.isEmpty) {
       for (var dir in _defaultWebDirs) {
         if (new Directory(dir).existsSync()) {
-          serveTargets.add(new _ServeTarget(dir, nextDefaultPort++));
+          serveTargets.add(new ServeTarget(dir, nextDefaultPort++));
         }
       }
     }
@@ -259,489 +188,37 @@ class _ServeOptions extends _SharedOptions {
     var buildDirs = _buildDirsFromOutputMap(outputMap)
       ..addAll(serveTargets.map((t) => t.dir));
 
-    return new _ServeOptions._(
-      hostName: argResults[_hostname] as String,
-      logRequests: argResults[_logRequests] as bool,
+    return new ServeOptions._(
+      hostName: argResults[hostnameOption] as String,
+      logRequests: argResults[logRequestsOption] as bool,
       serveTargets: serveTargets,
-      assumeTty: argResults[_assumeTty] as bool,
-      deleteFilesByDefault: argResults[_deleteFilesByDefault] as bool,
-      failOnSevere: argResults[_failOnSevere] as bool,
-      enableLowResourcesMode: argResults[_lowResourcesMode] as bool,
-      configKey: argResults[_config] as String,
+      assumeTty: argResults[assumeTtyOption] as bool,
+      deleteFilesByDefault: argResults[deleteFilesByDefaultOption] as bool,
+      failOnSevere: argResults[failOnSevereOption] as bool,
+      enableLowResourcesMode: argResults[lowResourcesModeOption] as bool,
+      configKey: argResults[configOption] as String,
       outputMap: _parseOutputMap(argResults),
-      trackPerformance: argResults[_trackPerformance] as bool,
-      skipBuildScriptCheck: argResults[_skipBuildScriptCheck] as bool,
-      verbose: argResults[_verbose] as bool,
+      trackPerformance: argResults[trackPerformanceOption] as bool,
+      skipBuildScriptCheck: argResults[skipBuildScriptCheckOption] as bool,
+      verbose: argResults[verboseOption] as bool,
       builderConfigOverrides:
-          _parseBuilderConfigOverrides(argResults[_define], rootPackage),
-      isReleaseBuild: argResults[_release] as bool,
+          _parseBuilderConfigOverrides(argResults[defineOption], rootPackage),
+      isReleaseBuild: argResults[releaseOption] as bool,
       buildDirs: buildDirs.toList(),
     );
   }
 }
 
 /// A target to serve, representing a directory and a port.
-class _ServeTarget {
+class ServeTarget {
   final String dir;
   final int port;
 
-  _ServeTarget(this.dir, this.port);
-}
-
-abstract class _BuildRunnerCommand extends Command<int> {
-  Logger get logger => new Logger(name);
-
-  List<BuilderApplication> get builderApplications =>
-      (runner as BuildCommandRunner).builderApplications;
-
-  PackageGraph get packageGraph => (runner as BuildCommandRunner).packageGraph;
-
-  _BuildRunnerCommand() {
-    _addBaseFlags();
-  }
-
-  void _addBaseFlags() {
-    argParser
-      ..addFlag(_assumeTty,
-          help: 'Enables colors and interactive input when the script does not'
-              ' appear to be running directly in a terminal, for instance when it'
-              ' is a subprocess',
-          negatable: true)
-      ..addFlag(_deleteFilesByDefault,
-          help:
-              'By default, the user will be prompted to delete any files which '
-              'already exist but were not known to be generated by this '
-              'specific build script.\n\n'
-              'Enabling this option skips the prompt and deletes the files. '
-              'This should typically be used in continues integration servers '
-              'and tests, but not otherwise.',
-          negatable: false,
-          defaultsTo: false)
-      ..addFlag(_lowResourcesMode,
-          help: 'Reduce the amount of memory consumed by the build process. '
-              'This will slow down builds but allow them to progress in '
-              'resource constrained environments.',
-          negatable: false,
-          defaultsTo: false)
-      ..addOption(_config,
-          help: 'Read `build.<name>.yaml` instead of the default `build.yaml`',
-          abbr: 'c')
-      ..addFlag(_failOnSevere,
-          help: 'Whether to consider the build a failure on an error logged.',
-          negatable: true,
-          defaultsTo: false)
-      ..addFlag(_trackPerformance,
-          help: r'Enables performance tracking and the /$perf page.',
-          negatable: true,
-          defaultsTo: false)
-      ..addFlag(_skipBuildScriptCheck,
-          help: r'Skip validation for the digests of files imported by the '
-              'build script.',
-          hide: true,
-          defaultsTo: false)
-      ..addMultiOption(_output,
-          help: 'A directory to write the result of a build to. Or a mapping '
-              'from a top-level directory in the package to the directory to '
-              'write a filtered build output to. For example "web:deploy".',
-          abbr: 'o')
-      ..addFlag(_verbose,
-          abbr: 'v',
-          defaultsTo: false,
-          negatable: false,
-          help: 'Enables verbose logging.')
-      ..addFlag(_release,
-          abbr: 'r',
-          defaultsTo: false,
-          negatable: true,
-          help: 'Build with release mode defaults for builders.')
-      ..addMultiOption(_define,
-          splitCommas: false,
-          help: 'Sets the global `options` config for a builder by key.');
-  }
-
-  /// Must be called inside [run] so that [argResults] is non-null.
-  ///
-  /// You may override this to return more specific options if desired, but they
-  /// must extend [_SharedOptions].
-  _SharedOptions _readOptions() {
-    return new _SharedOptions.fromParsedArgs(
-        argResults, argResults.rest, packageGraph.root.name, this);
-  }
-}
-
-/// A [Command] that does a single build and then exits.
-class _BuildCommand extends _BuildRunnerCommand {
-  @override
-  String get invocation => '${super.invocation} [directories]';
-
-  @override
-  String get name => 'build';
-
-  @override
-  String get description =>
-      'Performs a single build on the specified targets and then exits.';
-
-  @override
-  Future<int> run() async {
-    var options = _readOptions();
-    var result = await build(
-      builderApplications,
-      deleteFilesByDefault: options.deleteFilesByDefault,
-      enableLowResourcesMode: options.enableLowResourcesMode,
-      failOnSevere: options.failOnSevere,
-      configKey: options.configKey,
-      assumeTty: options.assumeTty,
-      outputMap: options.outputMap,
-      packageGraph: packageGraph,
-      verbose: options.verbose,
-      builderConfigOverrides: options.builderConfigOverrides,
-      isReleaseBuild: options.isReleaseBuild,
-      trackPerformance: options.trackPerformance,
-      skipBuildScriptCheck: options.skipBuildScriptCheck,
-      buildDirs: options.buildDirs,
-    );
-    if (result.status == BuildStatus.success) {
-      return ExitCode.success.code;
-    } else {
-      return result.failureType.exitCode;
-    }
-  }
-}
-
-/// A [Command] that watches the file system for updates and rebuilds as
-/// appropriate.
-class _WatchCommand extends _BuildRunnerCommand {
-  @override
-  String get invocation => '${super.invocation} [directories]';
-
-  @override
-  String get name => 'watch';
-
-  @override
-  String get description =>
-      'Builds the specified targets, watching the file system for updates and '
-      'rebuilding as appropriate.';
-
-  @override
-  Future<int> run() async {
-    var options = _readOptions();
-    var handler = await watch(
-      builderApplications,
-      deleteFilesByDefault: options.deleteFilesByDefault,
-      enableLowResourcesMode: options.enableLowResourcesMode,
-      failOnSevere: options.failOnSevere,
-      configKey: options.configKey,
-      assumeTty: options.assumeTty,
-      outputMap: options.outputMap,
-      packageGraph: packageGraph,
-      trackPerformance: options.trackPerformance,
-      skipBuildScriptCheck: options.skipBuildScriptCheck,
-      verbose: options.verbose,
-      builderConfigOverrides: options.builderConfigOverrides,
-      isReleaseBuild: options.isReleaseBuild,
-      buildDirs: options.buildDirs,
-    );
-    if (handler == null) return ExitCode.config.code;
-
-    await handler.currentBuild;
-    await handler.buildResults.drain();
-    return ExitCode.success.code;
-  }
-}
-
-/// Extends [_WatchCommand] with dev server functionality.
-class _ServeCommand extends _WatchCommand {
-  _ServeCommand() {
-    argParser
-      ..addOption(_hostname,
-          help: 'Specify the hostname to serve on', defaultsTo: 'localhost')
-      ..addFlag(_logRequests,
-          defaultsTo: false,
-          negatable: false,
-          help: 'Enables logging for each request to the server.');
-  }
-
-  @override
-  String get invocation => '${super.invocation} [<directory>[:<port>]]...';
-
-  @override
-  String get name => 'serve';
-
-  @override
-  String get description =>
-      'Runs a development server that serves the specified targets and runs '
-      'builds based on file system updates.';
-
-  @override
-  _ServeOptions _readOptions() => new _ServeOptions.fromParsedArgs(
-      argResults, argResults.rest, packageGraph.root.name, this);
-
-  @override
-  Future<int> run() async {
-    var options = _readOptions();
-    var handler = await watch(
-      builderApplications,
-      deleteFilesByDefault: options.deleteFilesByDefault,
-      enableLowResourcesMode: options.enableLowResourcesMode,
-      failOnSevere: options.failOnSevere,
-      configKey: options.configKey,
-      assumeTty: options.assumeTty,
-      outputMap: options.outputMap,
-      packageGraph: packageGraph,
-      trackPerformance: options.trackPerformance,
-      skipBuildScriptCheck: options.skipBuildScriptCheck,
-      verbose: options.verbose,
-      builderConfigOverrides: options.builderConfigOverrides,
-      isReleaseBuild: options.isReleaseBuild,
-      buildDirs: options.buildDirs,
-    );
-
-    if (handler == null) return ExitCode.config.code;
-
-    _ensureBuildWebCompilersDependency(packageGraph, logger);
-    var servers = await Future.wait(options.serveTargets
-        .map((target) => _startServer(options, target, handler)));
-    await handler.currentBuild;
-    // Warn if in serve mode with no servers.
-    if (options.serveTargets.isEmpty) {
-      logger.warning(
-          'Found no known web directories to serve, but running in `serve` '
-          'mode. You may expliclity provide a directory to serve with trailing '
-          'args in <dir>[:<port>] format.');
-    } else {
-      for (var target in options.serveTargets) {
-        stdout.writeln('Serving `${target.dir}` on '
-            'http://${options.hostName}:${target.port}');
-      }
-    }
-    await handler.buildResults.drain();
-    await Future.wait(servers.map((server) => server.close()));
-
-    return ExitCode.success.code;
-  }
-}
-
-Future<HttpServer> _startServer(
-    _ServeOptions options, _ServeTarget target, ServeHandler handler) async {
-  var server = await _bindServer(options, target);
-  serveRequests(
-      server, handler.handlerFor(target.dir, logRequests: options.logRequests));
-  return server;
-}
-
-Future<HttpServer> _bindServer(_ServeOptions options, _ServeTarget target) {
-  switch (options.hostName) {
-    case 'any':
-      // Listens on both IPv6 and IPv4
-      return HttpServer.bind(InternetAddress.anyIPv6, target.port);
-    case 'localhost':
-      return HttpMultiServer.loopback(target.port);
-    default:
-      return HttpServer.bind(options.hostName, target.port);
-  }
-}
-
-/// A [Command] that does a single build and then runs tests using the compiled
-/// assets.
-class _TestCommand extends _BuildRunnerCommand {
-  @override
-  String get invocation =>
-      '${super.invocation.replaceFirst('[arguments]', '[build-arguments]')} '
-      '[-- [test-arguments]]';
-
-  @override
-  String get name => 'test';
-
-  @override
-  String get description =>
-      'Performs a single build of the test directory only and then runs tests '
-      'using the compiled assets.';
-
-  @override
-  _SharedOptions _readOptions() {
-    // This command doesn't allow specifying directories to build, instead it
-    // always builds the `test` directory.
-    //
-    // Here we validate that [argResults.rest] is exactly equal to all the
-    // arguments after the `--`.
-    if (argResults.rest.isNotEmpty) {
-      void throwUsageException() {
-        throw new UsageException(
-            'The `test` command does not support positional args before the, '
-            '`--` separator, which should separate build args from test args.',
-            usage);
-      }
-
-      var separatorPos = argResults.arguments.indexOf('--');
-      if (separatorPos < 0) {
-        throwUsageException();
-      }
-      var expectedRest = argResults.arguments.skip(separatorPos + 1).toList();
-      if (argResults.rest.length != expectedRest.length) {
-        throwUsageException();
-      }
-      for (var i = 0; i < argResults.rest.length; i++) {
-        if (expectedRest[i] != argResults.rest[i]) {
-          throwUsageException();
-        }
-      }
-    }
-
-    return new _SharedOptions.fromParsedArgs(
-        argResults, ['test'], packageGraph.root.name, this);
-  }
-
-  @override
-  Future<int> run() async {
-    _SharedOptions options;
-    // We always run our tests in a temp dir.
-    var tempPath = Directory.systemTemp
-        .createTempSync('build_runner_test')
-        .absolute
-        .uri
-        .toFilePath();
-    try {
-      _ensureBuildTestDependency(packageGraph);
-      options = _readOptions();
-      var outputMap = options.outputMap ?? {};
-      outputMap.addAll({tempPath: null});
-      var result = await build(
-        builderApplications,
-        deleteFilesByDefault: options.deleteFilesByDefault,
-        enableLowResourcesMode: options.enableLowResourcesMode,
-        failOnSevere: options.failOnSevere,
-        configKey: options.configKey,
-        assumeTty: options.assumeTty,
-        outputMap: outputMap,
-        packageGraph: packageGraph,
-        trackPerformance: options.trackPerformance,
-        skipBuildScriptCheck: options.skipBuildScriptCheck,
-        verbose: options.verbose,
-        builderConfigOverrides: options.builderConfigOverrides,
-        isReleaseBuild: options.isReleaseBuild,
-        buildDirs: options.buildDirs,
-      );
-
-      if (result.status == BuildStatus.failure) {
-        stdout.writeln('Skipping tests due to build failure');
-        return result.failureType.exitCode;
-      }
-
-      var testExitCode = await _runTests(tempPath);
-      if (testExitCode != 0) {
-        // No need to log - should see failed tests in the console.
-        exitCode = testExitCode;
-      }
-      return testExitCode;
-    } on _BuildTestDependencyError catch (e) {
-      stdout.writeln(e);
-      return ExitCode.config.code;
-    } finally {
-      // Clean up the output dir.
-      await new Directory(tempPath).delete(recursive: true);
-    }
-  }
-
-  /// Runs tests using [precompiledPath] as the precompiled test directory.
-  Future<int> _runTests(String precompiledPath) async {
-    stdout.writeln('Running tests...\n');
-    var extraTestArgs = argResults.rest;
-    var testProcess = await Process.start(
-        _pubBinary,
-        [
-          'run',
-          'test',
-          '--precompiled',
-          precompiledPath,
-        ]..addAll(extraTestArgs),
-        mode: ProcessStartMode.inheritStdio);
-    return testProcess.exitCode;
-  }
-}
-
-class _CleanCommand extends Command<int> {
-  _CleanCommand();
-
-  @override
-  String get name => 'clean';
-
-  @override
-  String get description =>
-      'Cleans up output from previous builds. Does not clean up --output '
-      'directories.';
-
-  Logger get logger => new Logger(name);
-
-  @override
-  Future<int> run() async {
-    var logSubscription = Logger.root.onRecord.listen(stdIOLogListener);
-
-    logger.warning('Deleting cache and generated source files.\n'
-        'This shouldn\'t be necessary for most applications, unless you have '
-        'made intentional edits to generated files (i.e. for testing). '
-        'Consider filing a bug at '
-        'https://github.com/dart-lang/build/issues/new if you are using this '
-        'to work around an apparent (and reproducible) bug.');
-
-    await logTimedAsync(logger, 'Cleaning up source outputs', () async {
-      var assetGraphFile = new File(assetGraphPath);
-      if (!assetGraphFile.existsSync()) {
-        logger.warning(
-            'No asset graph found, skipping generated to source file cleanup');
-      } else {
-        var assetGraph =
-            new AssetGraph.deserialize(await assetGraphFile.readAsBytes());
-        var packageGraph = new PackageGraph.forThisPackage();
-        var writer = new FileBasedAssetWriter(packageGraph);
-        for (var id in assetGraph.outputs) {
-          if (id.package != packageGraph.root.name) continue;
-          var node = assetGraph.get(id) as GeneratedAssetNode;
-          if (node.wasOutput) {
-            // Note that this does a file.exists check in the root package and
-            // only tries to delete the file if it exists. This way we only
-            // actually delete to_source outputs, without reading in the build
-            // actions.
-            await writer.delete(id);
-          }
-        }
-      }
-    });
-
-    await logTimedAsync(logger, 'Cleaning up cache directory', () async {
-      var generatedDir = new Directory(cacheDir);
-      if (await generatedDir.exists()) {
-        await generatedDir.delete(recursive: true);
-      }
-    });
-
-    await logSubscription.cancel();
-
-    return 0;
-  }
+  ServeTarget(this.dir, this.port);
 }
 
 Set<String> _buildDirsFromOutputMap(Map<String, String> outputMap) =>
     outputMap.values.where((v) => v != null).toSet();
-
-void _ensureBuildTestDependency(PackageGraph packageGraph) {
-  if (!packageGraph.allPackages.containsKey('build_test')) {
-    throw new _BuildTestDependencyError();
-  }
-}
-
-void _ensureBuildWebCompilersDependency(PackageGraph packageGraph, Logger log) {
-  if (!packageGraph.allPackages.containsKey('build_web_compilers')) {
-    log.warning('''
-    Missing dev dependency on package:build_web_compilers, which is required to serve Dart compiled to JavaScript.
-
-    Please update your dev_dependencies section of your pubspec.yaml:
-
-    dev_dependencies:
-      build_runner: any
-      build_test: any
-      build_web_compilers: any''');
-  }
-}
 
 Map<String, Map<String, dynamic>> _parseBuilderConfigOverrides(
     dynamic parsedArg, String rootPackage) {
@@ -754,7 +231,7 @@ Map<String, Map<String, dynamic>> _parseBuilderConfigOverrides(
     if (parts.length < 3) {
       throw new ArgumentError.value(
           define,
-          _define,
+          defineOption,
           'Expected at least 2 `=` signs, should be of the format like '
           '$expectedFormat');
     } else if (parts.length > 3) {
@@ -784,16 +261,29 @@ Map<String, Map<String, dynamic>> _parseBuilderConfigOverrides(
   return builderConfigOverrides;
 }
 
-class _BuildTestDependencyError extends StateError {
-  _BuildTestDependencyError() : super('''
-Missing dev dependency on package:build_test, which is required to run tests.
-
-Please update your dev_dependencies section of your pubspec.yaml:
-
-  dev_dependencies:
-    build_runner: any
-    build_test: any
-    # If you need to run web tests, you will also need this dependency.
-    build_web_compilers: any
-''');
+/// Returns a map of output directory to root input directory to be used
+/// for merging.
+///
+/// Each output option is split on `:` where the first value is the
+/// root input directory and the second value output directory.
+/// If no delimeter is provided the root input directory will be null.
+Map<String, String> _parseOutputMap(ArgResults argResults) {
+  var outputs = argResults[outputOption] as List<String>;
+  if (outputs == null) return null;
+  var result = <String, String>{};
+  for (var option in argResults[outputOption] as List<String>) {
+    var split = option.split(':');
+    if (split.length == 1) {
+      var output = split.first;
+      result[output] = null;
+    } else if (split.length >= 2) {
+      var output = split.sublist(1).join(':');
+      var root = split.first;
+      if (root.contains('/')) {
+        throw 'Input root can not be nested: $option';
+      }
+      result[output] = split.first;
+    }
+  }
+  return result;
 }
